requirements.txt
# === Azure SDKs ===
azure-ai-ml>=1.23.0
azure-identity>=1.19.0
azure-storage-blob>=12.24.0
azure-mgmt-datafactory>=9.0.0
azure-mgmt-resource>=23.2.0

# === Azure OpenAI + GenAI ===
openai>=1.57.0
semantic-kernel>=1.18.0

# === ML / Data Science ===
scikit-learn>=1.5.0
xgboost>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.9.0
seaborn>=0.13.0
mlflow>=2.18.0
joblib>=1.4.0

# === Notebook / dev ===
jupyter>=1.1.0
ipykernel>=6.29.0
python-dotenv>=1.0.1

# === Quality ===
pytest>=8.3.0
black>=24.10.0
ruff>=0.8.0
